Move executeCommand
It is used from controllers rather than NewMantleBackupReconciler. No
functional changes.
