+++ internal/controller/util.go
@@ -2,7 +2,9 @@ package controller
 
 import (
 	"context"
+	"io"
 	"log/slog"
+	"os/exec"
 	"strings"
 
 	corev1 "k8s.io/api/core/v1"
@@ -49,3 +51,53 @@ func updateStatus(ctx context.Context, client client.Client, obj client.Object, 
 	}
 	return nil
 }
+
+func executeCommandImpl(command []string, input io.Reader) ([]byte, error) {
+	cmd := exec.Command(command[0], command[1:]...)
+	stdout, err := cmd.StdoutPipe()
+	if err != nil {
+		return nil, err
+	}
+	defer func() {
+		err := stdout.Close()
+		if err != nil {
+			logger.Error("failed to stdout.Close", "error", err)
+		}
+	}()
+
+	if input != nil {
+		stdin, err := cmd.StdinPipe()
+		if err != nil {
+			return nil, err
+		}
+		go func() {
+			defer func() {
+				err := stdin.Close()
+				if err != nil {
+					logger.Error("failed to stdin.Close", "error", err)
+				}
+			}()
+			if _, err = io.Copy(stdin, input); err != nil {
+				logger.Error("failed to io.Copy", "error", err)
+			}
+		}()
+	}
+
+	if err := cmd.Start(); err != nil {
+		return nil, err
+	}
+
+	r, err := io.ReadAll(stdout)
+	if err != nil {
+		return r, err
+	}
+
+	if err := cmd.Wait(); err != nil {
+		return r, err
+	}
+
+	return r, nil
+}
+
+// It will replaced on tests.
+var executeCommand = executeCommandImpl
